lendme/requirements.txt
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
black==23.12.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==42.0.2
dadata==21.10.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.8
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.1.0
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2023.12.1
filelock==3.13.1
flake8==6.1.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
identify==2.5.34
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.18
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.1
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.4
pytest-cov==4.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
gunicorn==22.0.0